fixed it

+++ ps2/include/qsort.h
@@ -3,7 +3,7 @@
 #include "stdbool.h"
 
 
-void my_qsort (void* base, size_t num, size_t size, int (*compar)(const void*,const void*));
+void my_qsort (void* base, size_t num, int size, int (*compar)(const void*,const void*));
 void* genericScan(void *X, size_t n, size_t l, void* (*oper)(void *x1, void *x2));
 void select_lower(void* base, size_t num, size_t size, void* pivot, int* swappable, int (*compar)(const void*, const void*));
